Refactor Add rental class
